+++ friends_server/src/main/kotlin/com/friends/security/service/AuthPasswordService.kt
@@ -2,11 +2,9 @@ package com.friends.security.service
 
 import com.friends.jwt.JwtService
 import com.friends.member.repository.MemberRepository
-import com.friends.security.securityException.DuplicateNewPasswordException
 import com.friends.security.securityException.EmailNotFoundException
-import com.friends.security.securityException.InvalidPasswordException
 import com.friends.security.securityException.InvalidTokenException
-import org.springframework.security.crypto.password.PasswordEncoder
+import com.friends.security.securityException.OAuth2ResetPasswordException
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
@@ -14,7 +12,7 @@ import org.springframework.transaction.annotation.Transactional
 class AuthPasswordService(
     private val jwtService: JwtService,
     private val memberRepository: MemberRepository,
-    private val passwordEncoder: PasswordEncoder,
+    private val authValidator: AuthValidator,
 ) {
     @Transactional
     fun resetPassword(
@@ -30,18 +28,13 @@ class AuthPasswordService(
         // 유저 존재 여부 확인
         val member = memberRepository.findByEmail(emailFromToken) ?: throw EmailNotFoundException()
 
-        // 패스워드 유효성 검사
-        // 적절한 비밀번호 패턴인지 검사
-        if (!validatePassword(newPassword)) {
-            throw InvalidPasswordException()
-        }
-        // 새 비밀번호와 기존 비밀번호가 같은지 검사
-        if (passwordEncoder.matches(newPassword, member.getPassword())) {
-            throw DuplicateNewPasswordException()
-        }
+        // OAuth2 회원의 경우 패스워드 변경 불가
+        val encryptedOldPassword = member.getPassword() ?: throw OAuth2ResetPasswordException()
 
-        // jwt 에서 email 을 추출하고 해당 email 을 가진 사용자의 비밀번호를 변경합니다.
-        member.updatePassword(passwordEncoder.encode(newPassword))
-        memberRepository.save(member)
+        // 패스워드 유효성 검사
+        authValidator.validateResetPassword(encryptedOldPassword, newPassword)
+
+        // 패스워드 변경
+        member.updatePassword(newPassword)
     }
 }
