+++ friends_server/src/main/kotlin/com/friends/security/service/AuthService.kt
@@ -1,31 +1,22 @@
 package com.friends.security.service
 
-import com.friends.category.CategoryNotFoundException
 import com.friends.category.repository.CategoryRepository
 import com.friends.config.AuthProperties
 import com.friends.image.S3ClientService
 import com.friends.jwt.AtRtService
 import com.friends.jwt.JwtService
 import com.friends.jwt.JwtType
-import com.friends.member.entity.Member
 import com.friends.member.entity.OAuth2Provider
 import com.friends.member.repository.MemberRepository
 import com.friends.oauth2.OAuth2Service
-import com.friends.profile.entity.Location
-import com.friends.profile.entity.Profile
-import com.friends.profile.entity.ProfileInterestTag
 import com.friends.profile.repository.ProfileInterestTagRepository
 import com.friends.profile.repository.ProfileRepository
 import com.friends.security.AtRtDto
 import com.friends.security.CheckEmailResponseDto
 import com.friends.security.CheckNicknameResponseDto
 import com.friends.security.OAuth2LoginDto
-import com.friends.security.RegisterRequestDto
 import com.friends.security.securityException.EmailDuplicateException
-import com.friends.security.securityException.InvalidNicknameException
-import com.friends.security.securityException.InvalidPasswordException
 import com.friends.security.securityException.InvalidRefreshTokenException
-import com.friends.security.securityException.InvalidTokenException
 import com.friends.security.userDetails.CustomUserDetails
 import org.springframework.security.authentication.AuthenticationManager
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
@@ -33,7 +24,6 @@ import org.springframework.security.core.authority.SimpleGrantedAuthority
 import org.springframework.security.crypto.password.PasswordEncoder
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
-import org.springframework.web.multipart.MultipartFile
 
 @Service
 class AuthService(
@@ -73,84 +63,6 @@ class AuthService(
         val memberId = atRtService.getMemberId(refreshToken)
         val authorities = atRtService.getAuthorities(refreshToken)
         return atRtService.createAtRt(memberId, authorities)
-    }
-
-    @Transactional
-    fun register(
-        registerRequestDto: RegisterRequestDto,
-        profileImage: MultipartFile?,
-    ) {
-        // emailAuthToken 검증
-        if (!jwtService.validate(registerRequestDto.authToken)) {
-            throw InvalidTokenException()
-        }
-
-        // Email-Password 회원가입인지, OAuth2 회원가입인지 확인
-        val type = jwtService.getClaim(registerRequestDto.authToken, "type", String::class.java)?.let { JwtType.valueOf(it) } ?: throw InvalidTokenException()
-        val user =
-            when (type) {
-                JwtType.EMAIL -> registerByEmail(registerRequestDto)
-                JwtType.OAUTH2 -> registerByOAuth2(registerRequestDto)
-            }
-
-        // 이메일이 중복되는지 검증
-        if (!validateEmail(user.email).isValid) {
-            throw EmailDuplicateException()
-        }
-
-        // 닉네임이 중복되는지 검증
-        if (!validateNickname(user.nickname).isValid) {
-            throw InvalidNicknameException()
-        }
-
-        // 유효성 검사를 통과한 멤버 저장
-        memberRepository.save(user)
-
-        // 프로필 생성
-        val profile =
-            Profile(
-                member = user,
-                imageUrl = profileImage?.let { s3ClientService.upload(it) },
-                gender = registerRequestDto.gender,
-                birth = registerRequestDto.birth,
-                location = registerRequestDto.location?.let { Location(it.latitude, it.longitude) },
-                selfDescription = registerRequestDto.selfDescription,
-                mbti = registerRequestDto.mbti,
-            )
-        profileRepository.save(profile)
-
-        // 프로필 관심사 태그 생성
-        profileInterestTagRepository.saveAll(
-            registerRequestDto.interestTag.map {
-                ProfileInterestTag(profile = profile, category = categoryRepository.findById(it).orElseThrow { CategoryNotFoundException() })
-            },
-        )
-    }
-
-    private fun registerByOAuth2(registerRequestDto: RegisterRequestDto): Member {
-        val email = jwtService.getClaim(registerRequestDto.authToken, "email", String::class.java) ?: throw InvalidTokenException()
-        val provider = jwtService.getClaim(registerRequestDto.authToken, "provider", String::class.java)?.let { OAuth2Provider.valueOf(it) } ?: throw InvalidTokenException()
-        val nickname = registerRequestDto.nickname
-        return Member.createUser(
-            nickname = nickname,
-            email = email,
-            oauth2Provider = provider,
-        )
-    }
-
-    private fun registerByEmail(registerRequestDto: RegisterRequestDto): Member {
-        val email = jwtService.getClaim(registerRequestDto.authToken, "email", String::class.java) ?: throw InvalidTokenException()
-        val nickname = registerRequestDto.nickname
-        val password = registerRequestDto.password
-        // 패스워드 유효성 검사
-        if (!Member.validatePassword(password!!)) {
-            throw InvalidPasswordException()
-        }
-        return Member.createUser(
-            nickname = nickname,
-            email = email,
-            password = passwordEncoder.encode(password),
-        )
     }
 
     fun validateNickname(nickname: String): CheckNicknameResponseDto {
